requirements.txt
fastapi
uvicorn
python-multipart
pymongo
langchain
langchain-openai
openai
pypdf
python-docx
tiktoken
